+++ src/components/Avatar/Avatar.module.scss
